+++ tests/puppeteer/campaigns/functional/BO/catalog/categories/03_categoriesBulkAction.js
@@ -7,8 +7,8 @@ const loginCommon = require('@commonTests/loginBO');
 const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
-const CategoriesPage = require('@pages/BO/categories');
-const AddCategoryPage = require('@pages/BO/addCategory');
+const CategoriesPage = require('@pages/BO/catalog/categories/categories');
+const AddCategoryPage = require('@pages/BO/catalog/categories/addCategory');
 const CategoryFaker = require('@data/faker/category');
 
 let browser;
