+++ tests/puppeteer/campaigns/functional/BO/customers/customers/03_customersBulkActionsInBO.js
@@ -7,8 +7,8 @@ const loginCommon = require('@commonTests/loginBO');
 const BOBasePage = require('@pages/BO/BObasePage');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
-const CustomersPage = require('@pages/BO/customers');
-const AddCustomerPage = require('@pages/BO/addCustomer');
+const CustomersPage = require('@pages/BO/customers/customers/customers');
+const AddCustomerPage = require('@pages/BO/customers/customers/addCustomer');
 const CustomerFaker = require('@data/faker/customer');
 
 let browser;
