+++ tests/puppeteer/campaigns/sanity/02_productsBO/04_deleteProduct.js
@@ -8,8 +8,8 @@ const loginCommon = require('@commonTests/loginBO');
 const LoginPage = require('@pages/BO/login');
 const DashboardPage = require('@pages/BO/dashboard');
 const BOBasePage = require('@pages/BO/BObasePage');
-const ProductsPage = require('@pages/BO/products');
-const AddProductPage = require('@pages/BO/addProduct');
+const ProductsPage = require('@pages/BO/catalog/products/products');
+const AddProductPage = require('@pages/BO/catalog/products/addProduct');
 const ProductFaker = require('@data/faker/product');
 
 let browser;
