+++ tests/puppeteer/pages/BO/catalog/categories/addCategory.js
