+++ tests/puppeteer/pages/BO/catalog/categories/categories.js
