+++ tests/puppeteer/pages/BO/catalog/products/addProduct.js
