+++ tests/puppeteer/pages/BO/catalog/products/products.js
