+++ tests/puppeteer/pages/BO/customers/customers/addCustomer.js
