+++ tests/puppeteer/pages/BO/customers/customers/viewCustomer.js
