+++ tests/puppeteer/pages/BO/international/taxes/taxes.js
