+++ tests/puppeteer/pages/BO/modules/moduleCatalog.js
