+++ tests/puppeteer/pages/BO/orders/orders/addOrder.js
@@ -1,7 +1,7 @@
 require('module-alias/register');
 const BOBasePage = require('@pages/BO/BObasePage');
 
-module.exports = class Order extends BOBasePage {
+module.exports = class AddOrder extends BOBasePage {
   constructor(page) {
     super(page);
