+++ tests/puppeteer/pages/BO/shopParameters/shopParamsGeneral.js
